requirements.txt
asgiref==3.5.2
autopep8==1.5.7
dj-database-url==0.5.0
Django==4.0.5
djangorestframework==3.13.1
django-extensions==3.1.5
gunicorn==20.1.0
Pillow==8.4.0
psycopg2-binary==2.9.3
python-decouple==3.6
python-dotenv==0.20.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0